feat: add functionality to reset forgoten password.
Works by creating a GUID saving it in the Database and sending an email
to the User with the GUID over the SendGrid API.

Then the user calls auth/forgot-password
And gives his Token, Email and his new Password.

Need to add a background Job to automatically delete expired Tokens and
a rate limiter for reseting, could use the Token for the rate limiting.